Few Changes for betterment of app.

+++ User_Interface_SplitMoney/controllers/homeController.js
@@ -18,15 +18,15 @@
 
                 //calculate total amount
                 angular.forEach($scope.cloneArray, function (val, key) {
-                    $scope.totalSpent = parseInt($scope.totalSpent) + parseInt(val.amount);
+                    $scope.totalSpent = (parseFloat($scope.totalSpent)) + (parseFloat(val.amount));
                 });
 
                 //calculate other amounts
                 angular.forEach($scope.cloneArray, function (val, key) {
                     var arrayLength = $scope.cloneArray.length;
-                    var equalDivision = parseInt($scope.totalSpent / arrayLength);
-                    val.pay = parseInt(val.amount - equalDivision) < 0 ? -parseInt(val.amount - equalDivision) : 0;
-                    val.receive = parseInt(val.amount - equalDivision) > 0 ? parseInt(val.amount - equalDivision) : 0;
+                    var equalDivision = parseFloat($scope.totalSpent / arrayLength);
+                    val.pay = (parseFloat(val.amount - equalDivision)) < 0 ? -(parseFloat(val.amount - equalDivision)) : 0;
+                    val.receive = (parseFloat(val.amount - equalDivision)) > 0 ? (parseFloat(val.amount - equalDivision)) : 0;
                     //$scope.userArray.push(userJson);
                 });
             };
